AI/Aspire-Full.Python/python-agents/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "aspire-agents"
version = "0.1.0"
description = "Thread-safe AI agents with Tensor Core compute for Python 3.15+ free-threaded runtime"
readme = "README.md"
requires-python = ">=3.15"
authors = [{ name = "Aspire Maintainers" }]
keywords = ["agents", "ai", "tensor", "gpu", "free-threaded", "python-3.15"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.15",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "openai>=1.51.0",
    "typer>=0.12.5",
    "rich>=13.8.0",
    "pydantic>=2.12.5",
    "PyYAML>=6.0.2",
    "httpx>=0.27.2",
    "tenacity>=9.0.0",
    "uvloop>=0.20.0; platform_system != 'Windows'",
    "torch>=2.5.1",
    "transformers>=4.37.0",
    "openai-agents[dapr]>=0.1.0",
    "mcp>=1.10.1",
    "websockets>=12.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "sounddevice>=0.5.0",
    "textual>=0.70.0",
    "requests>=2.31.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "playwright>=1.48.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.3",
    "mypy>=1.13.0",
    "types-requests>=2.31.0",
]
tracing = [
    "opentelemetry-sdk>=1.34.1",
    "opentelemetry-exporter-otlp-proto-http>=1.34.1",
    "opentelemetry-instrumentation-openai-v2>=2.1b0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
]

[project.scripts]
aspire-agents = "aspire_agents.cli:app"

[tool.uv]
managed = true
package = true
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "pytorch-nightly"
url = "https://download.pytorch.org/whl/nightly/cu126"
explicit = false

[tool.ruff]
target-version = "py314"  # Python 3.15t compatible (using py314 until ruff supports py315)
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["I", "E", "W", "F", "UP", "B", "SIM", "RUF"]  # Added UP (pyupgrade), B (bugbear), SIM (simplify), RUF (ruff-specific)
ignore = ["E203", "E501", "UP007"]  # UP007: Use X | Y instead of Union (already using)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
known-first-party = ["aspire_agents"]
force-single-line = false
lines-after-imports = 2

[tool.black]
line-length = 120
target-version = ["py313"]  # Python 3.15t compatible (using py313 until black supports py315)
preview = true

[tool.mypy]
python_version = "3.15"
files = "src"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
mypy_path = "src"
ignore_missing_imports = true
cache_dir = ".cache/mypy"


[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
]
